tests: Simplify setup after pytest 8.1 release
Since the release of pytest 8.1, some pytest options are no longer
needed and editable install can be used when running tests using tox.

As the --import-mode=importlib option is removed, this requires to
move the conftest.py file at the root of the repository to avoid
import errors when running tests.